\name{BeQTLR-package}
\alias{BeQTLR-package}
\alias{BeQTLR}
\docType{package}
\title{
\packageTitle{BeQTLR}
}
\description{
\packageDescription{BeQTLR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BeQTLR}
\packageIndices{BeQTLR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BeQTLR}

Maintainer: \packageMaintainer{BeQTLR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
